refactor: share one width-generic page engine between u32 and u64
The u32 and u64 codecs had near-identical encode_page/decode_page/best_bit
implementations differing only in element width. Extract a single generic
FastPForEngine<N, T> parameterized by a FastPForInt trait that abstracts the
width-specific pieces (bit width, exception bitmap word count, pack/unpack,
shifts). u32 keeps its hand-unrolled bit-packing kernels via trait dispatch,
so there is no performance regression; u64 uses the generic wide packer.

fastpfor.rs and fastpfor64.rs become thin wrappers over the engine. Net
-296 lines with the page algorithm now written once.

Verified byte-identical: u32 encode_compare fuzz (Rust==C++, 808K runs) and
u64 differential fuzz (619K runs) both clean, plus the full test suite.
