+++ src/rust/integer_compression/mod.rs
@@ -3,5 +3,6 @@ pub mod bitpacking_wide;
 pub mod bitunpacking;
 pub mod fastpfor;
 pub mod fastpfor64;
+pub mod fastpfor_engine;
 pub mod just_copy;
 pub mod variable_byte;
